User-to-user matching works :)
Added user-to-user (polling) communication. Once all NUMBER_OF_USERS
users are connected, random pairs are matched, and one of them is
randomly selected to make the first offer. By polling the server, they
can receive a new offer when it comes, or get notified that the
opponent accepted the deal.
Valuations are fixed for now, next commit should pull them from a
random distribution. After that, dealing with MySQL should be the next
step...