Updated instructions and sys.path issue fixed
When running the server from the terminal (as opposed to from PyCharm),
an ImportError would be thrown when trying to import any module(s) from
the
package BargainingExperiment. This was caused by sys.path not
containing BargainingExperiment's parent folder. A simple script has
been added to fix this issue, and now the server can be executed from
the terminal as well.
In addition, server installation instructions have been updated to
include latest changes (eg, use UCR's db instead of localhost).