Huge update: computer AI, allow users to quit before finishing...
Main changes:
 - Extended experiment settings available: role valuations, probability
of each valuation, starting role, match order (whether to play among
users or against the computer, and for how many rounds), etc.
 - Computer AI: implemented a "computer user" that plays against humans
(or replaces a human that decided to quit before the end of the
experiment).
 - Quit option: Due to IRB restrictions, allow users to quit at any
time. This also means that we need the computer AI to play for them (so
humans matched against them can still play).
 - Improved scheduler: ensures users are only matched once, and avoids
two sellers or two buyers facing each other.
 - Added a new user state: experiment complete. This state represents
the end of the whole experiment, and would show a summary of
points/profit obtained.
 - Restricted offers: Users are only allowed to choose from 2 single
values when making an offer (buyer's lower valuation or seller's higher
valuation), so no more sliding bars :(